requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pywin32>=306
PyYAML>=6.0.1
python-dotenv>=1.0.0
pytest>=7.4.0
dxcam>=0.3.0
